--- apps/ai-content-generator/src/components/app/dialog/common-generator/new-or-existing-text/NewOrExistingText.tsx
@@ -1,44 +1,0 @@
-import { Button, Flex } from '@contentful/f36-components';
-import { Dispatch } from 'react';
-import { GeneratorAction, GeneratorReducer } from '../generatorReducer';
-import { styles } from './NewOrExistingText.styles';
-
-interface Props {
-  isNewText: boolean;
-  dispatch: Dispatch<GeneratorReducer>;
-  isTitle?: boolean;
-}
-
-const NewOrExistingText = (props: Props) => {
-  const { isNewText, dispatch, isTitle } = props;
-
-  const changeTextSource = (isNew: boolean) => {
-    if (isNew === isNewText) {
-      return;
-    }
-
-    const type = isNew ? GeneratorAction.IS_NEW_TEXT : GeneratorAction.IS_NOT_NEW_TEXT;
-    dispatch({ type });
-  };
-
-  return (
-    <Flex flexGrow={1} flexDirection="column">
-      <Button
-        css={styles.button}
-        title="Input custom text"
-        isActive={isNewText}
-        onClick={() => changeTextSource(true)}>
-        {isTitle ? 'From prompt' : 'From new text'}
-      </Button>
-      <Button
-        title="Use existing text from entry"
-        css={styles.button}
-        isActive={!isNewText}
-        onClick={() => changeTextSource(false)}>
-        From field
-      </Button>
-    </Flex>
-  );
-};
-
-export default NewOrExistingText;
