+++ apps/ai-content-generator/src/configs/features/featureTypes.tsx
@@ -6,10 +6,6 @@ import { AIFeature } from './featureConfig';
 
 export type Prompt = (input: string, targetLocale: string) => string;
 
-export type FeatureComponentProps = {
-  isTitle?: boolean;
-};
-
 /**
  * Feature Items
  */
